Share Array resize reserve core rules

+++ src/runtime/object/xarray.h
@@ -112,8 +112,8 @@ XR_FUNC bool xr_array_has(XrArray *arr, XrValue value);
 XR_FUNC bool xr_array_is_empty(XrArray *arr);
 
 XR_FUNC void xr_array_fill(XrArray *arr, XrValue value, int64_t start, int64_t end);
-XR_FUNC bool xr_array_reserve(XrArray *arr, int32_t capacity);
-XR_FUNC bool xr_array_resize(XrArray *arr, int32_t length, XrValue fill);
+XR_FUNC bool xr_array_reserve(XrArray *arr, int64_t capacity);
+XR_FUNC bool xr_array_resize(XrArray *arr, int64_t length, XrValue fill);
 
 /* ====== Utility Methods ====== */
 
@@ -122,7 +122,7 @@ XR_FUNC XrArray *xr_array_copy(struct XrCoroutine *coro, XrArray *arr);
 /* ====== Internal Functions ====== */
 
 XR_FUNC void xr_array_grow(XrArray *arr);
-XR_FUNC void xr_array_ensure_capacity(XrArray *arr, int min_capacity);
+XR_FUNC void xr_array_ensure_capacity(XrArray *arr, int64_t min_capacity);
 
 /* ====== Slice Operations (zero-copy) ====== */
